feat: unified card search system and responsive menu bar
- Merged 7 redundant search components into UnifiedCardSearch
- Removed: EnhancedCardSearch, ScryfalInspiredSearch, ScryfalLikeAdvancedSearch, SyntaxAdvancedSearch, ScryfallSearchEngine, ScryfallSearchInterface, SearchIntegration
- Added comprehensive UnifiedCardSearch with all features from merged components
- Enhanced menu bar responsiveness with improved layout and spacing
- Added KONIVRER brand logo and optimized navigation for better visibility
- Created unified CSS styling system for search components
- Improved search performance with optimized indexing and debouncing
- Added comprehensive documentation and feature breakdown
